requirements.txt
fastapi==0.115.5
uvicorn[standard]==0.32.1
playwright==1.48.0
playwright-stealth==1.0.6
scrapling==0.2.8
cloudscraper==1.2.71
python-telegram-bot==21.7
requests==2.32.3
beautifulsoup4==4.12.3
pillow==11.0.0
aiofiles==24.1.0
python-multipart==0.0.20
websockets==14.1
pydantic==2.10.3
pydantic-settings==2.6.1
mitmproxy==10.1.6
selenium-wire==5.1.0
undetected-chromedriver==3.5.4
fake-useragent==1.4.0
python-anticaptcha==1.0.0
2captcha-python==1.2.1
pycryptodome==3.19.0
redis==5.0.1
celery==5.3.4
sqlalchemy==2.0.23
alembic==1.13.0
psutil==5.9.6
prometheus-client==0.19.0
# opencv-python-headless==4.8.1.78
# pytesseract==0.3.10
# speech-recognition removed (not available)
# pydub==0.25.1
# ffmpeg-python==0.2.0
numpy==1.26.2
pandas==2.1.4
lxml==4.9.3
aiohttp==3.9.1
httpx>=0.27.0
